+++ webpack.config.js
@@ -1,0 +1,9 @@
+const path = require('path');
+
+module.exports = {
+    entry: './src/app.ts',
+    output: {
+        filename: 'bundle.js',
+        path: path.resolve(__dirname, 'dist')
+    }
+}
